Add support for plotting the entries in the usercache
Add support for plotting the entries in the usercache.

This adds a new method to the plotting helper that reads the geojson from the
usercache instead of generating it directly.

Also create a new ipython notebook similar to end-to-end pipeline that uses the
new function to display trips for me and Tom.